Use repo2docker for building images

+++ hubploy/helm.py
@@ -28,16 +28,18 @@ def helm_upgrade(
     namespace,
     chart,
     config_files,
-    config_overrides
+    config_overrides,
+    version
 ):
     # Clear charts and do a helm dep up before installing
     # Clearing charts is important so we don't deploy charts that
     # have been removed from requirements.yaml
     # FIXME: verify if this is actually true
-    shutil.rmtree(os.path.join(chart, 'charts'), ignore_errors=True)
-    subprocess.check_call([
-        'helm', 'dep', 'up'
-    ], cwd=chart)
+    if os.path.exists(chart):
+        shutil.rmtree(os.path.join(chart, 'charts'), ignore_errors=True)
+        subprocess.check_call([
+            'helm', 'dep', 'up'
+        ], cwd=chart)
 
     cmd = [
         'helm',
@@ -46,9 +48,11 @@ def helm_upgrade(
         '--install',
         '--namespace', namespace,
         name, chart,
-    ] 
-    cmd += itertools.chain(*[['-f', cf] for cf in config_files]) 
-    cmd += itertools.chain(*[['--set', f'{k}={v}'] for k, v in config_overrides.items()])
+    ]
+    if version:
+        cmd += ['--version', version]
+    cmd += itertools.chain(*[['-f', cf] for cf in config_files])
+    cmd += itertools.chain(*[['--set', v] for v in config_overrides])
     subprocess.check_call(cmd)
 
 
@@ -57,6 +61,8 @@ def deploy(
     chart,
     environment,
     namespace=None,
+    config_overrides=None,
+    version=None,
 ):
     """
     Deploy a JupyterHub.
@@ -78,6 +84,9 @@ def deploy(
     `jupyterhub.singleuser.image.tag` will be automatically set to this image
     tag.
     """
+    if config_overrides is None:
+        config_overrides = []
+
     name = f'{deployment}-{environment}'
 
     if namespace is None:
@@ -94,20 +103,15 @@ def deploy(
             os.path.join('deployments', deployment, 'image')
         )
 
-        # FIXME: Make this more configurable
-        config_overrides = {
-            'jupyterhub.singleuser.image.tag': image_tag
-        }
-    else:
-        config_overrides = {}
-
+        config_overrides.append(f'jupyterhub.singleuser.image.tag={image_tag}')
 
     helm_upgrade(
         name,
         namespace,
         chart,
         config_files,
-        config_overrides
+        config_overrides,
+        version
     )
 
 
@@ -121,16 +125,23 @@ def main():
     )
     argparser.add_argument(
         'environment',
-        choices=['staging', 'prod']
+        choices=['develop', 'staging', 'prod']
     )
     argparser.add_argument(
         '--namespace',
         default=None
     )
+    argparser.add_argument(
+        '--set',
+        action='append',
+    )
+    argparser.add_argument(
+        '--version',
+    )
 
     args = argparser.parse_args()
 
-    deploy(args.deployment, args.chart, args.environment, args.namespace)
+    deploy(args.deployment, args.chart, args.environment, args.namespace, args.set, args.version)
 
 if __name__ == '__main__':
     main()
